# PlaylistChaser.Api/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "PlaylistChaser.Api",
    "version": "1.0"
  },
  "paths": {
    "/api/Account/register-user": {
      "post": {
        "tags": [
          "Account"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterUserModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterUserModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterUserModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Playlist/get-all-playlists": {
      "get": {
        "tags": [
          "Playlist"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Playlist"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Playlist"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Playlist"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Playlist/{id}": {
      "get": {
        "tags": [
          "Playlist"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Playlist"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Playlist"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Playlist"
                }
              }
            }
          }
        }
      }
    },
    "/api/Playlist/get-playlist-populated/{id}": {
      "get": {
        "tags": [
          "Playlist"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PlaylistPopulated"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaylistPopulated"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaylistPopulated"
                }
              }
            }
          }
        }
      }
    },
    "/api/Playlist/remove-playlist/{id}": {
      "put": {
        "tags": [
          "Playlist"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Playlist/remove-playlists": {
      "put": {
        "tags": [
          "Playlist"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            },
            "text/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            },
            "application/*+json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Playlist/get-thumbnail/{playlistId}": {
      "get": {
        "tags": [
          "Playlist"
        ],
        "parameters": [
          {
            "name": "playlistId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Playlist/sync-from-origin/{playlistId}": {
      "post": {
        "tags": [
          "Playlist"
        ],
        "parameters": [
          {
            "name": "playlistId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Playlist/sync-to-destinations/{playlistId}": {
      "post": {
        "tags": [
          "Playlist"
        ],
        "parameters": [
          {
            "name": "playlistId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/{sourceId}/Source/get-playlist/{url}": {
      "get": {
        "tags": [
          "Source"
        ],
        "parameters": [
          {
            "name": "url",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sourceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PlaylistInfo"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaylistInfo"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaylistInfo"
                }
              }
            }
          }
        }
      }
    },
    "/api/{sourceId}/Source/validate-playlist-url": {
      "post": {
        "tags": [
          "Source"
        ],
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sourceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "boolean"
                }
              },
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              },
              "text/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
    "/api/{sourceId}/Source/add-playlist": {
      "post": {
        "tags": [
          "Source"
        ],
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sourceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/{sourceId}/Source/GetValidSourceIdsList": {
      "get": {
        "tags": [
          "Source"
        ],
        "parameters": [
          {
            "name": "sourceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SourceId"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SourceId"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SourceId"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Playlist": {
        "required": [
          "channelName",
          "name",
          "originSourceId",
          "playlistTypeId",
          "userId"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "channelName": {
            "minLength": 1,
            "type": "string"
          },
          "thumbnailId": {
            "type": "integer",
            "format": "int32"
          },
          "thumbnail": {
            "$ref": "#/components/schemas/Thumbnail"
          },
          "playlistTypeId": {
            "$ref": "#/components/schemas/PlaylistTypes"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "originSourceId": {
            "$ref": "#/components/schemas/SourceId"
          },
          "userId": {
            "type": "integer",
            "format": "int32"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        },
        "additionalProperties": false
      },
      "PlaylistInfo": {
        "required": [
          "creatorName",
          "isMine",
          "lastSynced",
          "name",
          "playlistIdSource",
          "url"
        ],
        "type": "object",
        "properties": {
          "playlistId": {
            "type": "integer",
            "format": "int32"
          },
          "playlist": {
            "$ref": "#/components/schemas/Playlist"
          },
          "sourceId": {
            "$ref": "#/components/schemas/SourceId"
          },
          "playlistIdSource": {
            "minLength": 1,
            "type": "string"
          },
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "creatorName": {
            "minLength": 1,
            "type": "string"
          },
          "isMine": {
            "type": "boolean"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "url": {
            "minLength": 1,
            "type": "string"
          },
          "lastSynced": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "PlaylistPopulated": {
        "type": "object",
        "properties": {
          "playlist": {
            "$ref": "#/components/schemas/Playlist"
          },
          "songs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Song"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PlaylistTypes": {
        "enum": [
          1,
          2
        ],
        "type": "integer",
        "format": "int32"
      },
      "RegisterUserModel": {
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "minLength": 1,
            "type": "string",
            "format": "email"
          },
          "password": {
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "Song": {
        "required": [
          "artistName",
          "songName"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "songName": {
            "minLength": 1,
            "type": "string"
          },
          "artistName": {
            "minLength": 1,
            "type": "string"
          },
          "thumbnailId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SourceId": {
        "enum": [
          1,
          2
        ],
        "type": "integer",
        "format": "int32"
      },
      "Thumbnail": {
        "required": [
          "fileContents"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "fileContents": {
            "type": "string",
            "format": "byte"
          },
          "url": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "userName": {
            "type": "string",
            "nullable": true
          },
          "normalizedUserName": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "normalizedEmail": {
            "type": "string",
            "nullable": true
          },
          "emailConfirmed": {
            "type": "boolean"
          },
          "passwordHash": {
            "type": "string",
            "nullable": true
          },
          "securityStamp": {
            "type": "string",
            "nullable": true
          },
          "concurrencyStamp": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "phoneNumberConfirmed": {
            "type": "boolean"
          },
          "twoFactorEnabled": {
            "type": "boolean"
          },
          "lockoutEnd": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lockoutEnabled": {
            "type": "boolean"
          },
          "accessFailedCount": {
            "type": "integer",
            "format": "int32"
          },
          "dbUserName": {
            "type": "string",
            "nullable": true
          },
          "dbPassword": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}